+++ src/StateSmithTest/Processes/ICompilation.cs
@@ -1,0 +1,8 @@
+﻿#nullable enable
+
+namespace StateSmithTest.Processes;
+
+public interface ICompilation
+{
+    SimpleProcess Run(string runArgs = "");
+}