+++ src/StateSmithTest/Processes/SimpleProcessException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+#nullable enable
+
+namespace StateSmithTest.Processes;
+
+public class SimpleProcessException : InvalidOperationException
+{
+    public SimpleProcessException(string? message) : base(message)
+    {
+    }
+}