+++ src/StateSmithTest/spec/2/CSharp/Spec2TestsCSharp.cs
@@ -13,10 +13,11 @@ public class Spec2TestsCSharp : Spec2Tests, IClassFixture<SharedCompilationFixtu
         SimpleProcess process = new()
         {
             WorkingDirectory = SharedCompilationFixture.OutputDirectory,
-            CommandAndArgs = $"dotnet bin/Debug/net6.0/Spec2Test.dll {testEvents}" // like 5x faster!!! Tip from https://github.com/dotnet/sdk/issues/8697#issuecomment-327943066
-            //CommandText = $"dotnet run --no-build {testEvents}"   // slow compared to above
+            ProgramPath = "dotnet",
+            //Args = $"run --no-build {testEvents}"   // slow compared to below
+            Args = $"bin/Debug/net8.0/Spec2Test.dll {testEvents}" // 5x faster!!! Tip from https://github.com/dotnet/sdk/issues/8697#issuecomment-327943066
         };
-        BashRunner.RunCommand(process);
+        process.Run();
 
         string output = process.StdOutput;
         output = StringUtils.RemoveEverythingBefore(output, "\nIGNORE_OUTPUT_BEFORE_THIS\n").Trim();