+++ src/StateSmithTest/spec/2/JavaScript/SharedCompilationFixture.cs
@@ -29,9 +29,10 @@ public class SharedCompilationFixture
         process = new()
         {
             WorkingDirectory = OutputDirectory,
-            CommandAndArgs = "node --check Spec2Sm.js"
+            ProgramPath = "node",
+            Args = " --check Spec2Sm.js"
         };
-        BashRunner.RunCommand(process, timeoutMs: 10000);
+        process.Run(timeoutMs: 10000);
     }
 
     public class MyGlueFile : IRenderConfigJavaScript