+++ src/StateSmithTest/spec/2/JavaScript/Spec2TestsJavaScript.cs
@@ -13,9 +13,10 @@ public class Spec2TestsJavaScript : Spec2Tests, IClassFixture<SharedCompilationF
         SimpleProcess process = new()
         {
             WorkingDirectory = SharedCompilationFixture.OutputDirectory,
-            CommandAndArgs = $"node index.js {testEvents}"
+            ProgramPath = "node",
+            Args = $"index.js {testEvents}"
         };
-        BashRunner.RunCommand(process);
+        process.Run();
 
         string output = process.StdOutput;
         output = StringUtils.RemoveEverythingBefore(output, "\nIGNORE_OUTPUT_BEFORE_THIS\n").Trim();