+++ src/StateSmith.Cli/CHANGELOG.md
@@ -8,6 +8,10 @@ Test/interim releases are not documented here.
 
 ---
 
+## WIP
+
+---
+
 ## [0.8.1]
 ### Added
 - Main menu option to check for updates